After removing the `.ts` extensions that were passed along to the compiled JavaScript server 'dist' directory, the build command functioned but my external requests to the Yelp API did not function and the client's 'dist' folder / React application were not being presented to the user when running `npm run start` at the root level.
After testing, the deployment of the `server.js` file in the server 'dist' directory could not access the environmental variables for the `YELP_API_KEY` or the `NODE_ENV`. When these were referenced, these were returning `undefined`. To resolve this, I created a new `.env` file at the root level which is accessible by this `server.js` file. Admittedly, I'm not too sure why this functions because when using `npm run develop` at the root level the API server does not have access to the environmental variables. For this reason I had to have two `.env` files: one at the root level and one at the 'server' directory level

--------------------------------

For reference:

When testing for Render/production deployment:

> npm run build
> npm run start

- This will reference the `.env` file at the root level which has `NODE_ENV` === 'production' and the `YELP_API_KEY`.

--------------------------------

For reference:

When testing with concurrently for development where we deploy the React application server and the API server:

> npm run develop

- This will reference the `.env` file at the "server" directory level. We will use `MONGODB_URI` here at this level for seeding the remote MONGODB Atlas database. This will also contain the `YELP_API_KEY` environmental variable to make external API calls