In `App.tsx`:
- Imported `outlet` from `react-router-dom`
- Imported `Auth0Provider`
- Imported 'Navbar'

Passed the `Auth0Provider` as a wrapper to encompass the entire app. Imported and passed the 'Navbar' component, and below it rendered the 'Outlet' component which will connect to our router and render the selected element/component.

A handful of the tests and state code was removed from `App.tsx` to `Home.tsx`